Split dashboard into curated operations and performance views
