+++ app/dashboard/operations/page.tsx
@@ -1,0 +1,5 @@
+import CuratedDashboardPage from "../_components/CuratedDashboardPage";
+
+export default function OperationsDashboardPage() {
+  return <CuratedDashboardPage view="operations" />;
+}
