+++ app/dashboard/performance/page.tsx
@@ -1,0 +1,5 @@
+import CuratedDashboardPage from "../_components/CuratedDashboardPage";
+
+export default function PerformanceDashboardPage() {
+  return <CuratedDashboardPage view="performance" />;
+}
